[move]Upgrade move and change upgrade module strategy
