+++ abi/decoder/Cargo.toml
@@ -12,7 +12,7 @@ anyhow = "1.0.41"
 bcs = "0.1"
 bcs-ext = {path = "../../commons/bcs_ext"}
 hex = "0.4.3"
-move-binary-format = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
+move-binary-format = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
 ordinal = "0.3.1"
 schemars = {git = "https://github.com/starcoinorg/schemars", rev = "6972da92f4360e1779168bb3fe0274c521d324e2"}
 serde = {version = "1.0.130", features = ["derive", "rc"]}
