+++ commons/proptest-helpers/Cargo.toml
@@ -8,6 +8,6 @@ version = "1.12.4"
 
 [dependencies]
 crossbeam = "0.7.3"
-diem-proptest-helpers = {package = "diem-proptest-helpers", git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
+diem-proptest-helpers = {package = "diem-proptest-helpers", git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
 proptest = "1.0.0"
 proptest-derive = "0.3.0"
