+++ state/api/src/chain_state.rs
@@ -1,7 +1,7 @@
 // Copyright (c) The Starcoin Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use anyhow::{ensure, format_err, Result};
+use anyhow::{ensure, Result};
 use merkle_tree::{blob::Blob, proof::SparseMerkleProof, RawKey};
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
@@ -11,26 +11,16 @@ use starcoin_types::language_storage::StructTag;
 use starcoin_types::state_set::AccountStateSet;
 use starcoin_types::write_set::WriteSet;
 use starcoin_types::{
-    access_path::AccessPath,
-    account_address::AccountAddress,
-    account_config::{AccountResource, BalanceResource},
-    account_state::AccountState,
-    state_set::ChainStateSet,
+    access_path::AccessPath, account_address::AccountAddress, account_config::AccountResource,
+    account_state::AccountState, state_set::ChainStateSet,
 };
-use starcoin_vm_types::account_config::{genesis_address, G_STC_TOKEN_CODE};
 use starcoin_vm_types::genesis_config::ChainId;
-use starcoin_vm_types::language_storage::ModuleId;
-use starcoin_vm_types::on_chain_resource::dao::{Proposal, ProposalAction};
-use starcoin_vm_types::on_chain_resource::{
-    Epoch, EpochData, EpochInfo, GlobalTimeOnChain, Treasury,
-};
-use starcoin_vm_types::sips::SIP;
+use starcoin_vm_types::on_chain_resource::{Epoch, EpochInfo, GlobalTimeOnChain};
 use starcoin_vm_types::token::token_code::TokenCode;
-use starcoin_vm_types::token::token_info::TokenInfo;
 use starcoin_vm_types::{
     move_resource::MoveResource, on_chain_config::OnChainConfig, state_view::StateView,
 };
-use std::convert::{TryFrom, TryInto};
+use std::convert::TryFrom;
 
 #[derive(Debug, Default, Eq, PartialEq, Clone, Serialize, Deserialize)]
 pub struct StateProof {
@@ -154,150 +144,7 @@ pub trait ChainStateWriter {
     fn flush(&self) -> Result<()>;
 }
 
-impl<T: ?Sized> StateReaderExt for T where T: StateView {}
-
-pub trait StateReaderExt: StateView {
-    /// Get AccountResource by address
-    fn get_account_resource(&self, address: AccountAddress) -> Result<Option<AccountResource>> {
-        self.get_resource::<AccountResource>(address)
-    }
-
-    /// Get Resource by type R
-    fn get_resource<R>(&self, address: AccountAddress) -> Result<Option<R>>
-    where
-        R: MoveResource + DeserializeOwned,
-    {
-        let access_path = AccessPath::new(address, R::resource_path());
-        self.get_resource_by_access_path(access_path)
-    }
-
-    fn get_resource_by_access_path<R>(&self, access_path: AccessPath) -> Result<Option<R>>
-    where
-        R: MoveResource + DeserializeOwned,
-    {
-        let r = self.get(&access_path).and_then(|state| match state {
-            Some(state) => Ok(Some(bcs_ext::from_bytes::<R>(state.as_slice())?)),
-            None => Ok(None),
-        })?;
-        Ok(r)
-    }
-
-    fn get_sequence_number(&self, address: AccountAddress) -> Result<u64> {
-        self.get_account_resource(address)?
-            .map(|resource| resource.sequence_number())
-            .ok_or_else(|| format_err!("Can not find account by address:{}", address))
-    }
-
-    fn get_on_chain_config<C>(&self) -> Result<Option<C>>
-    where
-        C: OnChainConfig,
-        Self: Sized,
-    {
-        C::fetch_config(self)
-    }
-
-    fn get_balance(&self, address: AccountAddress) -> Result<Option<u128>> {
-        self.get_balance_by_token_code(address, G_STC_TOKEN_CODE.clone())
-    }
-
-    /// Get balance by address and coin type
-    fn get_balance_by_type(
-        &self,
-        address: AccountAddress,
-        type_tag: StructTag,
-    ) -> Result<Option<u128>> {
-        Ok(self
-            .get(&AccessPath::new(
-                address,
-                BalanceResource::access_path_for(type_tag),
-            ))
-            .and_then(|bytes| match bytes {
-                Some(bytes) => Ok(Some(bcs_ext::from_bytes::<BalanceResource>(
-                    bytes.as_slice(),
-                )?)),
-                None => Ok(None),
-            })?
-            .map(|resource| resource.token()))
-    }
-
-    fn get_balance_by_token_code(
-        &self,
-        address: AccountAddress,
-        token_code: TokenCode,
-    ) -> Result<Option<u128>> {
-        self.get_balance_by_type(address, token_code.try_into()?)
-    }
-
-    fn get_epoch(&self) -> Result<Epoch> {
-        self.get_resource::<Epoch>(genesis_address())?
-            .ok_or_else(|| format_err!("Epoch is none."))
-    }
-
-    fn get_epoch_info(&self) -> Result<EpochInfo> {
-        let epoch = self
-            .get_resource::<Epoch>(genesis_address())?
-            .ok_or_else(|| format_err!("Epoch is none."))?;
-
-        let epoch_data = self
-            .get_resource::<EpochData>(genesis_address())?
-            .ok_or_else(|| format_err!("Epoch is none."))?;
-
-        Ok(EpochInfo::new(epoch, epoch_data))
-    }
-
-    fn get_timestamp(&self) -> Result<GlobalTimeOnChain> {
-        self.get_resource(genesis_address())?
-            .ok_or_else(|| format_err!("Timestamp resource should exist."))
-    }
-
-    fn get_chain_id(&self) -> Result<ChainId> {
-        self.get_resource::<ChainId>(genesis_address())?
-            .ok_or_else(|| format_err!("ChainId resource should exist at genesis address. "))
-    }
-
-    fn get_code(&self, module_id: ModuleId) -> Result<Option<Vec<u8>>> {
-        self.get(&AccessPath::from(&module_id))
-    }
-
-    /// Check the sip is activated. if the sip module exist, think it is activated.
-    fn is_activated(&self, sip: SIP) -> Result<bool> {
-        self.get_code(sip.module_id()).map(|code| code.is_some())
-    }
-
-    fn get_token_info(&self, token_code: TokenCode) -> Result<Option<TokenInfo>> {
-        let type_tag = token_code.try_into()?;
-        let access_path = TokenInfo::resource_path_for(type_tag);
-        self.get_resource_by_access_path(access_path)
-    }
-
-    fn get_stc_info(&self) -> Result<Option<TokenInfo>> {
-        self.get_token_info(G_STC_TOKEN_CODE.clone())
-    }
-
-    fn get_treasury(&self, token_code: TokenCode) -> Result<Option<Treasury>> {
-        let access_path = Treasury::resource_path_for(token_code.try_into()?);
-        self.get_resource_by_access_path(access_path)
-    }
-
-    fn get_stc_treasury(&self) -> Result<Option<Treasury>> {
-        self.get_treasury(G_STC_TOKEN_CODE.clone())
-    }
-
-    fn get_proposal<A>(&self, token_code: TokenCode) -> Result<Option<Proposal<A>>>
-    where
-        A: ProposalAction + DeserializeOwned,
-    {
-        let access_path = Proposal::<A>::resource_path_for(token_code.try_into()?);
-        self.get_resource_by_access_path(access_path)
-    }
-
-    fn get_stc_proposal<A>(&self) -> Result<Option<Proposal<A>>>
-    where
-        A: ProposalAction + DeserializeOwned,
-    {
-        self.get_proposal(G_STC_TOKEN_CODE.clone())
-    }
-}
+pub use starcoin_vm_types::state_view::StateReaderExt;
 
 /// `AccountStateReader` is a helper struct for read account state.
 pub struct AccountStateReader<'a, Reader> {
