+++ vm/compiler/Cargo.toml
@@ -8,9 +8,9 @@ version = "1.12.4"
 
 [dependencies]
 anyhow = "1.0.41"
-move-binary-format = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
-move-command-line-common = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
-move-compiler = {package = "move-compiler", git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
+move-binary-format = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
+move-command-line-common = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
+move-compiler = {package = "move-compiler", git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
 once_cell = "1.13.0"
 petgraph = "0.5.1"
 regex = {version = "1.6.0", default-features = false, features = [
