+++ vm/starcoin-transactional-test-harness/Cargo.toml
@@ -29,11 +29,11 @@ jsonrpc-core-client = {version = "18", features = [
   "arbitrary_precision",
 ]}
 log = {version = "0.4.16"}
-move-binary-format = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
-move-command-line-common = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
-move-compiler = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
-move-core-types = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
-move-transactional-test-runner = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
+move-binary-format = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
+move-command-line-common = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
+move-compiler = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
+move-core-types = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
+move-transactional-test-runner = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
 once_cell = "~1"
 serde = {version = "1"}
 serde_json = {version = "1"}
@@ -46,7 +46,7 @@ futures = "0.3.12"
 jsonrpc-core = {version = "18", features = ["arbitrary_precision"]}
 jsonrpc-derive = "18"
 jsonrpc-http-server = "18"
-move-resource-viewer = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
+move-resource-viewer = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
 starcoin-abi-decoder = {path = "../../abi/decoder"}
 starcoin-accumulator = {package = "starcoin-accumulator", path = "../../commons/accumulator"}
 starcoin-chain-api = {path = "../../chain/api"}
