+++ vm/transaction-builder-generator/Cargo.toml
@@ -19,7 +19,7 @@ serde_yaml = "0.8.26"
 textwrap = "0.14.0"
 
 bcs = "0.1.3"
-move-core-types = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
+move-core-types = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
 starcoin-vm-types = {path = "../types"}
 
 [dev-dependencies]
