+++ vm/vm-runtime/Cargo.toml
@@ -10,11 +10,11 @@ version = "1.12.4"
 anyhow = "1.0.41"
 bcs-ext = {package = "bcs-ext", path = "../../commons/bcs_ext"}
 crypto = {package = "starcoin-crypto", git = "https://github.com/starcoinorg/starcoin-crypto", rev = "d871dfb4216f034ee334a575926c101574d9d6dc"}
-move-core-types = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
+move-core-types = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
 once_cell = "1.13.0"
 tracing = "0.1.34"
 
-move-vm-runtime = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749"}
+move-vm-runtime = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5"}
 num_enum = "0.5.7"
 rand = "0.8.5"
 #prometheus = { version = "0.13.1", default_features = false }
@@ -25,7 +25,7 @@ starcoin-natives = {path = "../natives"}
 starcoin-types = {path = "../../types"}
 starcoin-vm-types = {path = "../types"}
 
-move-stdlib = {git = "https://github.com/starcoinorg/move", rev = "ae90a6d2453fcceebfc546d572ca4f91da29b749", optional = true}
+move-stdlib = {git = "https://github.com/starcoinorg/move", rev = "c6dc63d3da2bfc625073536001a50b0d2fd093f5", optional = true}
 starcoin-metrics = {path = "../../commons/metrics", optional = true}
 
 [dev-dependencies]
